+++ tools/rendercli/rendercli.cpp
@@ -421,12 +421,16 @@ namespace {
         << unsignedValue(batching, "frontierCompactionCandidateDepths")
         << " frontier_compaction_candidate_samples="
         << unsignedValue(batching, "frontierCompactionCandidateSamples")
+        << " frontier_compaction_candidate_packed_ray_bytes="
+        << unsignedValue(batching, "frontierCompactionCandidatePackedRayBytes")
         << " frontier_compaction_candidate_fraction="
         << batching.value("frontierCompactionCandidateSampleFraction").toDouble()
         << " frontier_largest_compaction_candidate_depth="
         << unsignedValue(batching, "frontierLargestCompactionCandidateDepth")
         << " frontier_largest_compaction_candidate_samples="
         << unsignedValue(batching, "frontierLargestCompactionCandidateSamples")
+        << " frontier_largest_compaction_candidate_packed_ray_bytes="
+        << unsignedValue(batching, "frontierLargestCompactionCandidatePackedRayBytes")
         << " frontier_largest_compaction_candidate_fraction="
         << batching.value("frontierLargestCompactionCandidateSampleFraction").toDouble()
         << " frontier_hit_rays=" << unsignedArraySum(frontierHits)